Make sweeping changes to whole file. Can now provide correct answers.
This is pretty close to a finished product, though it is not polished.

I figured out how to define a process of elimination in prolog, finally.

At the moment, the queries to get correct answers have to have very
specific formats:

  ownership: person(VAR1), alien(VAR2), owns(VAR1, VAR2).
  powers: alien(VAR1), power(VAR2), has_power(VAR1, VAR2).

The reason that these forms are necessary is that the facts make *heavy*
use of the goal (!, fail), which cuts away backtracking in the current
predicate being invoked (the current predicate which is the goal). The
previous two things must be there to maintain a point to backtrack to
which the cut operator (`!`) cannot touch.

Changes to be committed:
      modified:   facts.prolog